% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpCalcResponse.R
\name{bdpCalcResponse}
\alias{bdpCalcResponse}
\title{Bayesian Deming regression (informative Prior) with Stan}
\usage{
bdpCalcResponse(stanRegr, Xval, ci = 0.95, ...)
}
\arguments{
\item{stanRegr}{bdpreg object}

\item{Xval}{Reference method data}

\item{ci}{Probability for the HDI credibility interval. Default 0.95.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
no return
}
\description{
Bayesian Deming regression (informative Prior) with Stan
}
