% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpExtract.R
\name{bdpExtract}
\alias{bdpExtract}
\title{Extract relevant data from the Bayesian Deming regression}
\usage{
bdpExtract(stanRegr, X, Y, ErrorRatio = 1)
}
\arguments{
\item{stanRegr}{bdpreg object}

\item{X}{Reference method data}

\item{Y}{Testing method data}

\item{ErrorRatio}{Deming variance ratio. Default = 1.}
}
\value{
A data frame with extracted data
}
\description{
Extract relevant data from the Bayesian Deming regression
}
