% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPairs.R
\name{bdpPairs}
\alias{bdpPairs}
\title{Pairs plot of the full posterior predictors}
\usage{
bdpPairs(bdpreg, ...)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{...}{Arguments passed to \code{rstan::pairs}.}
}
\value{
Pairs plot of the \code{stanfit} object
}
\description{
Pairs plot of the full posterior predictors
}
