% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPlot.R
\name{bdpPlot}
\alias{bdpPlot}
\title{Plot Bayesian Deming Regression}
\usage{
bdpPlot(stanRegr, X, Y, ci = 0.95)
}
\arguments{
\item{stanRegr}{Rstan bdpreg object to plot}

\item{X}{Reference method data}

\item{Y}{Testing method data}

\item{ci}{Probability for the HDI credibility interval. Default 0.95}
}
\value{
no return
}
\description{
Plot Bayesian Deming Regression
}
