% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxEllipses.R
\name{bdpPlotBE}
\alias{bdpPlotBE}
\title{Plot regression posterior pairs with CI Box and MD ellipses}
\usage{
bdpPlotBE(bdpreg, cov.method = "MCD", ci = 0.95)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{cov.method}{rrcov covariance method ("SDe", "MCD", or "Classical"). Default MCD.}

\item{ci}{Probability for the HDI credibility interval. Default 0.95.}
}
\value{
no return
}
\description{
Plot regression posterior pairs with CI Box and MD ellipses
}
