% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPlotResiduals.R
\name{bdpPlotResiduals}
\alias{bdpPlotResiduals}
\title{Plot studentized residuals from the Bayesian Deming regression}
\usage{
bdpPlotResiduals(bdpreg, legend.pos = "bottomleft", legend.horizon = TRUE, ...)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{legend.pos}{Legend position (e.g. \code{bottomleft}, \code{bottomright}, \code{topright}). Default \code{bottomleft}}

\item{legend.horizon}{Legend \code{horiz} boolean parameter. Default TRUE.}

\item{...}{Arguments passed to \code{plot} (e.g. \code{xlim}, \code{xlab}, \code{main})}
}
\value{
no return
}
\description{
Plot studentized residuals from the Bayesian Deming regression
}
