% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPrint.R
\name{bdpPrint}
\alias{bdpPrint}
\title{Print summary of sampled data}
\usage{
bdpPrint(bdpreg, digits_summary = 4, ...)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{digits_summary}{number of digits for the results}

\item{...}{Arguments passed to \code{rstan::print}}
}
\value{
Print of the \code{stanfit} object
}
\description{
Print summary of sampled data
}
