% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpSummary.R
\name{bdpSummary}
\alias{bdpSummary}
\title{Print summary of sampled data}
\usage{
bdpSummary(bdpreg, ...)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{...}{Arguments passed to \code{rstan::summary}}
}
\value{
Summary list of the \code{stanfit} object
}
\description{
Print summary of sampled data
}
