% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpTraceplot.R
\name{bdpTraceplot}
\alias{bdpTraceplot}
\title{Plot a traceplot of the sampled chains}
\usage{
bdpTraceplot(bdpreg, ...)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{...}{Arguments passed to \code{rstan::traceplot} (e.g. iter, chains).}
}
\value{
Traceplot of the \code{stanfit} object
}
\description{
Plot a traceplot of the sampled chains
}
