\name{glycHem}
    \alias{glycHem}
    \docType{data}
    \title{Method comparison of glycated hemoglobin measurement}
    \description{
       This data gives the glycated hemoglobin measured in 20 patients.
    }
    \usage{data(glycHem)}
    \format{A data frame with 20 observations on the following 2 variables.}
    \keyword{datasets}
