% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBoxEllipses.R
\name{plotBE}
\alias{plotBE}
\title{Plot regression posterior pairs}
\usage{
plotBE(stanRegr, cov.method = "MCD", ci = 0.95)
}
\arguments{
\item{stanRegr}{Rstan rstanbdp object}

\item{cov.method}{rrcov covariance method ("SDe", "MCD", or "Classical"). Default MCD.}

\item{ci}{Probability for the HDI credibility interval. Default 0.95.}
}
\value{
no return
}
\description{
Plot regression posterior pairs
}
