% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanbdp-package.R
\name{rstanbdp-package}
\alias{rstanbdp-package}
\alias{rstanbdp}
\title{The 'rstanbdp' package.}
\description{
Bayesian Deming regression for Method Comparison
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.5. https://mc-stan.org
_PATH
}
