% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpCalcResponse.R
\name{bdpCalcResponse}
\alias{bdpCalcResponse}
\title{Plot the calculated Y response with CI from the full Bayesian posterior distribution}
\usage{
bdpCalcResponse(bdpreg, Xval, ci = 0.95, ...)
}
\arguments{
\item{bdpreg}{bdpreg object}

\item{Xval}{Reference method data}

\item{ci}{Probability for the HDI credibility interval. Default 0.95.}

\item{...}{Arguments passed to \code{hist} (e.g. \code{breaks}, \code{xlim}, ...).}
}
\value{
no return
}
\description{
Plot the calculated Y response with CI from the full Bayesian posterior distribution
}
