% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpExtract.R
\name{bdpExtract}
\alias{bdpExtract}
\title{Extract relevant data from the Bayesian Deming regression}
\usage{
bdpExtract(bdpreg)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}
}
\value{
A data frame with extracted data
}
\description{
Extract relevant data from the Bayesian Deming regression
}
