% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPlot.R
\name{bdpPlot}
\alias{bdpPlot}
\title{Plot Bayesian Deming regression and the confidence intervals from the full posterior distribution}
\usage{
bdpPlot(bdpreg, ci = 0.95)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}

\item{ci}{Probability for the HDI credibility interval. Default 0.95}
}
\value{
no return
}
\description{
Plot Bayesian Deming regression and the confidence intervals from the full posterior distribution
}
