% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdpPlotResiduals.R
\name{bdpPlotResiduals}
\alias{bdpPlotResiduals}
\title{Plot studentized residuals from the Bayesian Deming regression}
\usage{
bdpPlotResiduals(bdpreg)
}
\arguments{
\item{bdpreg}{bdpreg object created with bdpreg}
}
\value{
no return
}
\description{
Plot studentized residuals from the Bayesian Deming regression
}
